API-33: Fixed commit issues.
