--- auth-service/src/test/java/com/sparkystudios/traklibrary/authentication/service/scheduler/AuthTaskSchedulerTest.java
@@ -1,95 +1,0 @@
-package com.sparkystudios.traklibrary.authentication.service.scheduler;
-
-import com.sparkystudios.traklibrary.authentication.domain.User;
-import com.sparkystudios.traklibrary.authentication.repository.UserRepository;
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.ExtendWith;
-import org.mockito.ArgumentMatchers;
-import org.mockito.InjectMocks;
-import org.mockito.Mock;
-import org.mockito.Mockito;
-import org.mockito.junit.jupiter.MockitoExtension;
-
-import java.util.Collections;
-
-@ExtendWith(MockitoExtension.class)
-class AuthTaskSchedulerTest {
-
-    @Mock
-    private UserRepository userRepository;
-
-    @InjectMocks
-    private AuthTaskScheduler authTaskScheduler;
-
-    @Test
-    void removeExpiredVerificationCodesScheduledTask_withNoUsers_doesntMakeChanges() {
-        // Arrange
-        Mockito.when(userRepository.findByVerifiedIsFalseAndVerificationExpiryDateBefore(ArgumentMatchers.any()))
-                .thenReturn(Collections.emptyList());
-
-        // Act
-        authTaskScheduler.removeExpiredVerificationCodesScheduledTask();
-
-        // Assert
-        Mockito.verify(userRepository, Mockito.never())
-                .save(ArgumentMatchers.any());
-    }
-
-    @Test
-    void removeExpiredVerificationCodesScheduledTask_withUsers_removesVerificationCodeAndExpiryDateAndSaves() {
-        // Arrange
-        User user = Mockito.spy(User.class);
-
-        Mockito.when(userRepository.findByVerifiedIsFalseAndVerificationExpiryDateBefore(ArgumentMatchers.any()))
-                .thenReturn(Collections.singletonList(user));
-
-        // Act
-        authTaskScheduler.removeExpiredVerificationCodesScheduledTask();
-
-        // Assert
-        Mockito.verify(user, Mockito.atMostOnce())
-                .setVerificationExpiryDate(null);
-
-        Mockito.verify(user, Mockito.atMostOnce())
-                .setVerificationCode(null);
-
-        Mockito.verify(userRepository, Mockito.atMostOnce())
-                .save(ArgumentMatchers.any());
-    }
-
-    @Test
-    void removeExpiredRecoveryTokensScheduledTask_withNoUsers_doesntMakeChanges() {
-        // Arrange
-        Mockito.when(userRepository.findByRecoveryTokenExpiryDateBefore(ArgumentMatchers.any()))
-                .thenReturn(Collections.emptyList());
-
-        // Act
-        authTaskScheduler.removeExpiredRecoveryTokensScheduledTask();
-
-        // Assert
-        Mockito.verify(userRepository, Mockito.never())
-                .save(ArgumentMatchers.any());
-    }
-
-    @Test
-    void removeExpiredRecoveryTokensScheduledTask_withUsers_removesRecoveryTokenAndExpiryDateAndSaves() {
-        // Arrange
-        User user = Mockito.spy(User.class);
-
-        Mockito.when(userRepository.findByRecoveryTokenExpiryDateBefore(ArgumentMatchers.any()))
-                .thenReturn(Collections.singletonList(user));
-
-        // Act
-        authTaskScheduler.removeExpiredRecoveryTokensScheduledTask();
-
-        // Assert
-        Mockito.verify(user, Mockito.atMostOnce())
-                .setRecoveryTokenExpiryDate(null);
-
-        Mockito.verify(user, Mockito.atMostOnce())
-                .setRecoveryToken(null);
-
-        Mockito.verify(userRepository, Mockito.atMostOnce())
-                .save(ArgumentMatchers.any());
-    }
-}
